gitattributes
I wish it weren't true, but a cross-platform code-base (with cross-platform devs): it's practically necessary to use gitattributes file(s).
Code files are especially tricky text files: with whitespace both significant and insignificant, both horizontal and vertical.  Telling git a file is text just handles EOLs, but EOLs are most likely to change without 'active' editing by the dev.
But neither do I like git to auto-detect which files are text as this is usually followed by "oh, btw" binaries.  I like to use gitattributes as a means to inform the contents of the repo without removing the save-every-byte (binary) default.